data preparation (PKAD+pkPDB compatible) excluding partial charges
HI Kloppel,

Here is the code to import the data from the manually downloaded PKAD (experimental) csv and pkPDB* (PYPKA) csv and turns it into the mini batches we discussed. However, in the mini_batches function, it only does one loop because downloading all of the PDBs from pkPDB, unlike from PKAD (the database I built this on and we were working with in the meeting and has since been modified to work with the pkPDB data, too). But, this loop works when it is allowed to loop through every row of the input dataframe, it was tested on the data from PKAD.

*I had to manually modify this file name in computer (.csv.gz --> .csv)