attempt 2 (866 structures minus unknown # of pdbs that failed.hopefully 200 epochs)
60/40 split, which was the same for attempt 1. hydrogens

Currently running this code for 200 epochs, and it is currently on epoch 12 and looks like it's gonna plateau without ever exploding (pictured). but, I have seen before that   means as a target function (as opposed to sums) behaves better- but does not necessarily perform (accuracy, flexibility) better-let's see. 

here, i increased MHA's "complexity value", for lack of a better word at the moment, from 1 in egnn run 1 (fail), a change which was done simply for minimal model/occams razor purposes and never once brought me success, back to 22, as it was in previous successes.

I made dim 1 (from 2), further decreased learning rate, made the target function from a mean-sum mixture to a single mean, and I shuffled my data. i added further security against data prep errors which could (although unlikely, as I've seen that the model is robust against noise) result in model disillusionment. 

importantly, each run means a redistribution of which pdbs are in validation and which are in test set, something I do not intend to keep. Something I did see produce changes back when I had a pretty homogeneous set of only 20 structures (specifically, the performance was sensitive to whether "10gs", a complicated and unique structure unlike the rest, was in the validation or training set, as one would expect).