data prep + 140 pdbs through the NHR, one without fixed metal parsing and one with.
the model runs should be comparable enough re: same input data. the model is the same. the experiment is reproducible, anyways. the nhr outputs are in the text files as well as the model codes, in "results" of PKParse. with 500 atoms per hood and all the same size hoods, no hydrogens. needs an rcsb gzipped pdb as reference pdb, and a modeled pdb, which picky pkegnn model seems to like, btw (unpictured).  this parser does not support parsing of pdbs with multiple occupancy, and will produce fallicous data if the insertion codes have not been corrected by a pdbmending tool such as pdbfixer. This would not support an amber LEAP file and only supports pdbs with chain values. it is really better to input only pdbs which go by the standard format used by anything using openMM, pdbfixer, probably MDAnalysis and Biopython, the latter which I believe uses OpenMM to save like many do, or with good probabality and a bit of luck, a PDB straight from the RCSB can work as the "modeled PDB", which is the parsed PDB for pkegnn data. the RCSB reference PDB is strictly for sequence alignment.

This parser uses AMBER atom naming convention, standard residues and their notations, and PKAI github's intrinsic PKA values to calculate shift values from pkpdb (unpictured) csv.
the model is randomly chosen in the sense that it was simply the last one that was used on the last version of my data prep with absent disulfide cysteines, a model choice (hyperparameter choice) sitting in between experiment and convenience. it should, however, be representative

There should finally be no missing structural data, except hydrogens and H20. the included ligands are in cofactors py dictionary in utils.py and ligands.npz in /data/.

This is a binary parser and will not accept ungzipped data. it is fine to simply bit encode the lines and should be implemented, there is probably a one liner for it. but i refrain from user functionality like that before it has generated 121K x ~100 structures (in other words, this code minimizes if statements as it is in Development and Efficiency Mode).