Please open a terminal in the folder where these 4 py files are saved and run train.py.
This model is two models with three different aggregation schemes across the neighbors, and two different aggregation schemes across the features. It can be used to study the model and easily do ablation studies or change a tunable parameter. More later in another more final branch, because e.g. I will be adding a nonlearnable aggregation across features later.

Note this wont run out of box due to the fact that the inputs is not as far away on your computer as mine. otherwise I dont see why it should not.